move out applyEdit
